test: add basetest configurations
